requirements.txt
# local package
-e .

# external requirements
setuptools
numpy

pyzmq
pandas
pyqt6
pyqtgraph